// src/modules/organization/domain/enums/employment-transition-reason.enum.ts
export enum EmploymentTransitionReason {
  NEW_HIRE = 'NEW_HIRE',
  PROMOTION = 'PROMOTION',
  TRANSFER = 'TRANSFER',
  EMPLOYMENT_TYPE_CHANGE = 'EMPLOYMENT_TYPE_CHANGE',
  DESIGNATION_CHANGE = 'DESIGNATION_CHANGE',
  TEAM_CHANGE = 'TEAM_CHANGE',
  DEPARTMENT_CHANGE = 'DEPARTMENT_CHANGE',
  BRANCH_CHANGE = 'BRANCH_CHANGE',
  RETURN_FROM_LEAVE = 'RETURN_FROM_LEAVE',
  MARKETPLACE_CONVERSION = 'MARKETPLACE_CONVERSION',
  REHIRE = 'REHIRE',
  TERMINATION = 'TERMINATION'
}